Move `create_handoff_tool()` to `chat_config.py`
The `create_handoff_tool()` function can be used in multiple chatbots,
therefore it was moved to the `chat_config.py` file to avoid code repetition.
