+++ scripts/07_langgraph_structured_routing.py
@@ -1,33 +1,12 @@
 from dotenv import load_dotenv
 from enum import Enum
-from langchain_core.tools import InjectedToolCallId
-from langgraph.graph import MessagesState, StateGraph, START
+from langgraph.graph import StateGraph, START
 from langgraph.graph.state import CompiledStateGraph
-from langgraph.prebuilt import InjectedState
-from langgraph.types import Command
 from pydantic import BaseModel, Field
-from typing import Annotated
 
 from chat_config import *
 from chat_history import ChatHistory
 from db import init_db
-
-
-def create_handoff_tool(*, agent_name: str, description: str | None = None) -> BaseTool:
-    name = f'transfer_to_{agent_name}'
-    description = description or f'Ask {agent_name} for help.'
-
-    @tool(name, description=description)
-    def handoff_tool(state: Annotated[MessagesState, InjectedState], tool_call_id: Annotated[str, InjectedToolCallId]) -> Command:
-        tool_message = {
-            'role': 'tool',
-            'content': f'{text_colors["violet2"]}Successfully transfered to {agent_name}.',
-            'name': name,
-            'tool_call_id': tool_call_id
-        }
-        return Command(goto=agent_name, update={**state, 'messages': state['messages'] + [tool_message]}, graph=Command.PARENT)
-
-    return handoff_tool
 
 
 def create_agents(model: BaseChatModel) -> list[CompiledGraph]:
